requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2        
Django==4.2.7
django-rest-framework==0.1.0     
djangorestframework==3.14.0      
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1      
google-auth-oauthlib==1.1.0      
google-cloud-api-keys==0.5.4     
google-cloud-core==2.3.3
google-cloud-storage==2.13.0     
google-crc32c==1.5.0
google-resumable-media==2.6.0    
googleapis-common-protos==1.61.0 
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
proto-plus==1.22.3
protobuf==4.25.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
